dropping v2 in variable name
