+++ src/main/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha/BUILD.bazel
@@ -8,12 +8,20 @@ package(default_visibility = [
 ])
 
 kt_jvm_library(
+    name = "resource_key",
+    srcs = glob(["*Key.kt"]) + ["IdVariable.kt"],
+    deps = [
+        "//src/main/kotlin/org/wfanet/measurement/common",
+    ],
+)
+
+kt_jvm_library(
     name = "requisition_fulfillment_service",
     srcs = ["RequisitionFulfillmentService.kt"],
     deps = [
+        ":resource_key",
         "//imports/kotlin/kotlinx/coroutines:core",
         "//src/main/kotlin/org/wfanet/measurement/common/grpc",
-        "//src/main/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha/utils:resource_key",
         "//src/main/kotlin/org/wfanet/measurement/duchy/storage:metric_value_store",
         "//src/main/kotlin/org/wfanet/measurement/system/v1alpha:resource_key",
         "//src/main/proto/wfa/measurement/api/v2alpha:requisition_fulfillment_service_kt_jvm_grpc",
