+++ src/main/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha/ProtocolConfigKey.kt
@@ -1,0 +1,36 @@
+// Copyright 2021 The Cross-Media Measurement Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.wfanet.measurement.duchy.service.api.v2alpha
+
+import org.wfanet.measurement.common.ResourceNameParser
+
+private val parser = ResourceNameParser("protocolConfigs/{protocol_config}")
+
+/** [ResourceKey] of a ProtocolConfig. */
+data class ProtocolConfigKey(val protocolConfigId: String) : ResourceKey {
+  override fun toName(): String {
+    return parser.assembleName(mapOf(IdVariable.PROTOCOL_CONFIG to protocolConfigId))
+  }
+
+  companion object {
+    val defaultValue = ProtocolConfigKey("")
+
+    fun fromName(resourceName: String): ProtocolConfigKey? {
+      return parser.parseIdVars(resourceName)?.let {
+        ProtocolConfigKey(it.getValue(IdVariable.PROTOCOL_CONFIG))
+      }
+    }
+  }
+}
