+++ src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha/DataProviderKey.kt
@@ -12,24 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.duchy.service.api.v2alpha.utils
+package org.wfanet.measurement.kingdom.service.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
 
-private val parser = ResourceNameParser("protocolConfigs/{protocol_config}")
+private val parser = ResourceNameParser("dataProviders/{data_provider}")
 
-/** [ResourceKey] of a ProtocolConfig. */
-data class ProtocolConfigKey(val protocolConfigId: String) : ResourceKey {
+/** [DataProviderKey] of a Data Provider. */
+data class DataProviderKey(val dataProviderId: String) : ResourceKey {
   override fun toName(): String {
-    return parser.assembleName(mapOf(IdVariable.PROTOCOL_CONFIG to protocolConfigId))
+    return parser.assembleName(mapOf(IdVariable.DATA_PROVIDER to dataProviderId))
   }
 
   companion object {
-    val defaultValue = ProtocolConfigKey("")
+    val defaultValue = DataProviderKey("")
 
-    fun fromName(resourceName: String): ProtocolConfigKey? {
+    fun fromName(resourceName: String): DataProviderKey? {
       return parser.parseIdVars(resourceName)?.let {
-        ProtocolConfigKey(it.getValue(IdVariable.PROTOCOL_CONFIG))
+        DataProviderKey(it.getValue(IdVariable.DATA_PROVIDER))
       }
     }
   }
