+++ src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha/ExchangeStepAttemptKey.kt
@@ -12,40 +12,44 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.kingdom.service.api.v2alpha.utils
+package org.wfanet.measurement.kingdom.service.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
 
 private val parser =
   ResourceNameParser(
-    "recurringExchanges/{recurring_exchange}/exchanges/{exchange}/steps/{exchange_step}"
+    "recurringExchanges/{recurring_exchange}/exchanges/{exchange}/steps/{exchange_step}/" +
+      "attempts/{exchange_step_attempt}"
   )
 
-/** [ExchangeStepKey] of an Exchange Step. */
-data class ExchangeStepKey(
+/** [ExchangeStepAttemptKey] of an Exchange Step Attempt. */
+data class ExchangeStepAttemptKey(
   val recurringExchangeId: String,
   val exchangeId: String,
-  val exchangeStepId: String
+  val exchangeStepId: String,
+  val exchangeStepAttemptId: String
 ) : ResourceKey {
   override fun toName(): String {
     return parser.assembleName(
       mapOf(
         IdVariable.RECURRING_EXCHANGE to recurringExchangeId,
         IdVariable.EXCHANGE to exchangeId,
-        IdVariable.EXCHANGE_STEP to exchangeStepId
+        IdVariable.EXCHANGE_STEP to exchangeStepId,
+        IdVariable.EXCHANGE_STEP_ATTEMPT to exchangeStepAttemptId
       )
     )
   }
 
   companion object {
-    val defaultValue = ExchangeStepKey("", "", "")
+    val defaultValue = ExchangeStepAttemptKey("", "", "", "")
 
-    fun fromName(resourceName: String): ExchangeStepKey? {
+    fun fromName(resourceName: String): ExchangeStepAttemptKey? {
       return parser.parseIdVars(resourceName)?.let {
-        ExchangeStepKey(
+        ExchangeStepAttemptKey(
           it.getValue(IdVariable.RECURRING_EXCHANGE),
           it.getValue(IdVariable.EXCHANGE),
-          it.getValue(IdVariable.EXCHANGE_STEP)
+          it.getValue(IdVariable.EXCHANGE_STEP),
+          it.getValue(IdVariable.EXCHANGE_STEP_ATTEMPT)
         )
       }
     }
