--- src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha/utils/BUILD.bazel
@@ -1,11 +1,0 @@
-load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_library")
-
-package(default_visibility = ["//visibility:public"])
-
-kt_jvm_library(
-    name = "resource_key",
-    srcs = glob(["*Key.kt"]) + ["IdVariable.kt"],
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/common",
-    ],
-)
