+++ src/test/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha/RequisitionFulfillmentServiceTest.kt
@@ -44,7 +44,6 @@ import org.wfanet.measurement.common.flatten
 import org.wfanet.measurement.common.grpc.testing.GrpcTestServerRule
 import org.wfanet.measurement.common.testing.chainRulesSequentially
 import org.wfanet.measurement.common.testing.verifyProtoArgument
-import org.wfanet.measurement.duchy.service.api.v2alpha.utils.RequisitionKey as RequisitionKeyV2
 import org.wfanet.measurement.duchy.storage.MetricValueStore
 import org.wfanet.measurement.internal.duchy.ComputationToken
 import org.wfanet.measurement.internal.duchy.ComputationsGrpcKt.ComputationsCoroutineImplBase
@@ -55,7 +54,7 @@ import org.wfanet.measurement.storage.StorageClient
 import org.wfanet.measurement.storage.filesystem.FileSystemStorageClient
 import org.wfanet.measurement.storage.testing.BlobSubject.Companion.assertThat
 import org.wfanet.measurement.system.v1alpha.FulfillRequisitionRequest as SystemFulfillRequisitionRequest
-import org.wfanet.measurement.system.v1alpha.RequisitionKey
+import org.wfanet.measurement.system.v1alpha.RequisitionKey as SystemRequisitionKey
 import org.wfanet.measurement.system.v1alpha.RequisitionsGrpcKt.RequisitionsCoroutineImplBase
 import org.wfanet.measurement.system.v1alpha.RequisitionsGrpcKt.RequisitionsCoroutineStub
 
@@ -68,13 +67,13 @@ private val SIGNATURE = ByteString.copyFromUtf8("a signature")
 private val HEADER =
   FulfillRequisitionRequest.Header.newBuilder()
     .apply {
-      name = RequisitionKeyV2(DATA_PROVIDER_ID, REQUISITION_ID).toName()
+      name = RequisitionKey(DATA_PROVIDER_ID, REQUISITION_ID).toName()
       dataProviderParticipationSignature = SIGNATURE
     }
     .build()
 private val FULFILLED_RESPONSE =
   FulfillRequisitionResponse.newBuilder().apply { state = Requisition.State.FULFILLED }.build()
-private val REQUISITION_KEY = RequisitionKey(COMPUTATION_ID, REQUISITION_ID)
+private val REQUISITION_KEY = SystemRequisitionKey(COMPUTATION_ID, REQUISITION_ID)
 
 /** Test for [RequisitionFulfillmentService]. */
 @RunWith(JUnit4::class)
